ml/requirements.txt
flask
pandas
scikit-learn
joblib
pdfplumber
python-docx
requests